// src/modules/client-adm/usecase/find-client/find-client.usecase.dto.ts
export interface FindClientInputDto {
    id: string
}

export interface FindClientOutputDto {
    id: string;
    name: string;
    email: string;
    street: string;
    number: string;
    complement: string;
    city: string;
    state: string;
    zipCode: string;
    document: string;
    createdAt: Date;
    updatedAt: Date;
}